pyproject.toml
[project]
name = "code-auditor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "celery[redis]>=5.6.3",
    "fastapi>=0.136.0",
    "openai>=2.32.0",
    "passlib[bcrypt]>=1.7.4",
    "prometheus-client>=0.25.0",
    "psycopg2-binary>=2.9.11",
    "pydantic[email]>=2.13.3",
    "pydantic-settings>=2.14.0",
    "pygithub>=2.9.1",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "redis>=6.4.0",
    "semgrep>=1.160.0",
    "slowapi>=0.1.9",
    "sqlalchemy[asyncio]>=2.0.49",
    "structlog>=25.5.0",
    "uvicorn[standard]>=0.44.0",
    "bcrypt==4.0.1",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.22.1",
    "httpx>=0.28.1",
    "mypy>=1.20.1",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.15.11",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=src --cov-report=term-missing --cov-fail-under=79"